pyproject.toml
[tool.poetry]
name = "base-django-dev-scaffold"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "base_django_dev_scaffold"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.7"
pandas = "^1.5.3"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
mysqlclient = "^2.1.1"
django-comment-migrate = "^0.1.7"
django-log-request-id = "^2.1.0"
colorlog = "^6.7.0"
django-simpleui = "^2023.3.1"
djangorestframework-simplejwt = "^5.2.2"
django-simple-captcha = "^0.5.17"


[[tool.poetry.source]]
name = "ali"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
secondary = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
